Crate pallet_inv4

Pallet INV4

Config

Call

Pallet

Overview

This pallet handles advanced virtual multisigs (internally called cores).

Lower level implementation details of this pallet’s calls are contained in separate modules, each of them containing their own docs.

Pallet Functions

create_core - Create a new core

token_mint - Mint the core’s voting token to a target (called by a core origin)

token_burn - Burn the core’s voting token from a target (called by a core origin)

operate_multisig - Create a new multisig proposal, auto-executing if caller passes execution threshold requirements

vote_multisig - Vote on an existing multisig proposal, auto-executing if caller puts vote tally past execution threshold requirements

withdraw_vote_multisig - Remove caller’s vote from an existing multisig proposal

cancel_multisig_proposal - Cancel an existing multisig proposal (called by a core origin)

set_parameters - Change core parameters incl. voting thresholds and token freeze state (called by a core origin)

Re-exports

pub use account_derivation::CoreAccountDerivation;

pub use weights::WeightInfo;

pub use pallet::*;

Modules

account_derivation

Core Account Derivation.

dispatch 🔒

Dispatches calls internally, charging fees to the multisig account.

fee_handling

MultisigFeeHandler trait.

inv4_core

Core creation and internal management.

lookup 🔒

Custom account lookup implementation.

multisig

Multisig Operations.

origin

Custom Multisig Origin (INV4Origin).

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

voting

Voting Mechanism.

weights

Autogenerated weights for pallet_inv4

Structs

INV4Lookup

StaticLookup implementor using MultiAddress::Index for looking up cores by id.