Module pallet_inv4::multisig

Multisig Operations.

Overview

Handles the core actions within an already established multisig.

Core functionalities:

Minting/Burning voting tokens to existing and new members.

Handling proposal votes.

Dispatching approved proposals when both support and approval meet/exceed their minimum required thresholds.

Canceling proposals.

Structs

MultisigOperation

Details of a multisig operation.

Constants

MAX_SIZE

Maximum size of call we store is 50kb.

Type Aliases

BoundedCallBytes

MultisigOperationOf