Function pallet_inv4::pallet::dispatchables::set_parameters

pub fn set_parameters<T: Config>(
    metadata: Option<BoundedVec<u8, T::MaxMetadata>>,
    minimum_support: Option<Perbill>,
    required_approval: Option<Perbill>,
    frozen_tokens: Option<bool>
)

Change core parameters incl. voting thresholds and token freeze state (called by a core origin)

metadata: Arbitrary byte vec to be attached to the core info

minimum_support: Minimum amount of positive votes out of total token supply required to approve a proposal

required_approval: Minimum amount of positive votes out of current positive + negative votes required to approve a proposal

frozen_tokens: Wheter or not the core’s voting token should be transferable by the holders

NOTE: This function is an automatically generated, doc only, uncallable stub. designed to document the set_parameters variant of the pallet Call enum. You should not attempt to call this function directly.