Module pallet_inv4::voting

Voting Mechanism.

Overview

This module provides a weighted voting Tally implementation used for managing the multisig’s proposals. Members each have a balance in voting tokens and this balance differentiate their voting power as every vote utilizes the entire power of the said member. This empowers decision-making where certain members possess greater influence.

Structs

Tally

Aggregated votes for an ongoing poll by members of a core.

Enums

Vote

Represents a proposal vote within a multisig context.

Traits

CustomPolling

Type Aliases

Core

VoteRecord

Type alias for Vote with BalanceOf.

Votes