Struct pallet_inv4::voting::Tally

pub struct Tally<T: Config> {
    pub ayes: Votes<T>,
    pub nays: Votes<T>,
    pub records: BoundedBTreeMap<T::AccountId, Vote<Votes<T>>, T::MaxCallers>,
    dummy: PhantomData<T>,
}

Aggregated votes for an ongoing poll by members of a core.

Fields
ayes: Votes<T>nays: Votes<T>records: BoundedBTreeMap<T::AccountId, Vote<Votes<T>>, T::MaxCallers>dummy: PhantomData<T>
Implementations

impl<T: Config> Tally<T>

pub fn from_parts( ayes: Votes<T>, nays: Votes<T>, records: BoundedBTreeMap<T::AccountId, Vote<Votes<T>>, T::MaxCallers> ) -> Self

Allows for building a Tally manually.

pub fn process_vote( &mut self, account: T::AccountId, maybe_vote: Option<Vote<Votes<T>>> ) -> Result<Vote<Votes<T>>, DispatchError>

Check if a vote is valid and add the member’s total voting token balance to the tally.