Module pallet_inv4::fee_handling

MultisigFeeHandler trait.

Overview

Defines how transaction fees are charged to the multisig account. This trait requires proper runtime implementation to allow the usage of native or non-native assets.

Enums

FeeAsset

Represents the asset to be used by the multisig for paying transaction fees.

FeeAssetNegativeImbalance

Represents a potential negative asset balance incurred during fee payment operations within a multisig context.

Traits

MultisigFeeHandler

Fee handler trait.