Module pallet_inv4::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

cancel_multisig_proposal

Cancel an existing multisig proposal (called by a core origin)

create_core

Create a new core

operate_multisig

Create a new multisig proposal, auto-executing if caller passes execution threshold requirements Fees are calculated using the length of the metadata and the call The proposed call’s weight is used internally to charge the multisig instead of the user proposing the call

set_parameters

Change core parameters incl. voting thresholds and token freeze state (called by a core origin)

token_burn

Burn the core’s voting token from a target (called by a core origin)

token_mint

Mint the core’s voting token to a target (called by a core origin)

vote_multisig

Vote on an existing multisig proposal, auto-executing if caller puts vote tally past execution threshold requirements

withdraw_vote_multisig

Remove caller’s vote from an existing multisig proposal